requirements.txt
flask>=2.0
pandas>=1.5
openpyxl>=3.0
requests>=2.25
python-Levenshtein